# budgetFlow-ui/src/app/services/models/category-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CategoryRequest {
  categoryType?: 'EXPENSE' | 'INCOME';
  hexColor?: string;
  id?: number;
  name?: string;
  parentId?: number;
}