# budgetFlow-ui/src/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OpenApi specification - Patryk",
    "description": "OpenApi documentation  for Spring security",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "Patryk",
      "url": "https://google.com",
      "email": "contatct@patryk.com"
    },
    "license": {
      "name": "License name",
      "url": "https://google.pl"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Local ENV"
    },
    {
      "url": "https://testsuper.com/api/v1",
      "description": "Prod ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/expenses/income": {
      "get": {
        "tags": [
          "Income"
        ],
        "operationId": "getAllIncomeByUser",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IncomeResponse"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Income"
        ],
        "operationId": "updateIncome",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncomeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncomeResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Income"
        ],
        "operationId": "addIncome",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncomeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncomeResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Income"
        ],
        "operationId": "deleteIncome",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/expenses/statistics": {
      "post": {
        "tags": [
          "Expense"
        ],
        "operationId": "getStatisticsByMonth",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatisticsByMonthRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseForStatistics"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/expenses/income/statistics": {
      "post": {
        "tags": [
          "Income"
        ],
        "operationId": "getStatisticsByMonth_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatisticsByMonthRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseForStatistics"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/expenses/income/incomeByMonth": {
      "post": {
        "tags": [
          "Income"
        ],
        "operationId": "getIncomeByUserByMonth",
        "parameters": [
          {
            "name": "inGroup",
            "in": "query",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatisticsByMonthRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IncomeResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/expenses/income/getSumOfIncomes": {
      "post": {
        "tags": [
          "Income"
        ],
        "operationId": "getSumOfIncomesByMonth",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatisticsByMonthRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SumResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/expenses/category": {
      "post": {
        "tags": [
          "Category"
        ],
        "operationId": "saveCategory",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/expenses/category/getCategoriesByType": {
      "post": {
        "tags": [
          "Category"
        ],
        "operationId": "getAllCategory",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/expenses/category/editCategory": {
      "post": {
        "tags": [
          "Category"
        ],
        "operationId": "editCategory",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/expenses/category/{categoryId}": {
      "delete": {
        "tags": [
          "Category"
        ],
        "operationId": "deleteCategory",
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CategoryRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "hexColor": {
            "type": "string"
          },
          "parentId": {
            "type": "integer",
            "format": "int64"
          },
          "categoryType": {
            "type": "string",
            "enum": [
              "EXPENSE",
              "INCOME"
            ]
          }
        }
      },
...
      "CategoryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "createdByUserId": {
            "type": "integer",
            "format": "int64"
          },
          "hexColor": {
            "type": "string"
          },
          "parentId": {
            "type": "integer",
            "format": "int64"
          },
          "subCategories": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "categoryType": {
            "type": "string",
            "enum": [
              "EXPENSE",
              "INCOME"
            ]
          }
        }
      },
...
      "IncomeRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "categoryRequest": {
            "$ref": "#/components/schemas/CategoryRequest"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "incomeDate": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "IncomeResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "incomeDate": {
            "type": "string",
            "format": "date"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "categoryResponse": {
            "$ref": "#/components/schemas/CategoryResponse"
          }
        },
        "additionalProperties": true
      },
      "StatisticsByMonthRequest": {
        "type": "object",
        "properties": {
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          }
        }
      },
      "ResponseForStatistics": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "averageValue": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "SumResponse": {
        "type": "object",
        "properties": {
          "sum": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth description",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}